+++ include/modules/dynamodb/DynamoDbTableList.h
@@ -46,6 +46,12 @@ public:
     ~DynamoDbTableList() override;
 
     /**
+     * @brief Clear the page content
+     */
+    void ClearContent() override {
+    }
+
+    /**
      * @brief Load page content
      */
     void LoadContent() override;