+++ include/modules/kms/KMSKeyList.h
@@ -47,6 +47,12 @@ public:
     ~KMSKeyList() override;
 
     /**
+     * @brief Clear the page content
+     */
+    void ClearContent() override {
+    }
+
+    /**
      * @brief Load page content
      */
     void LoadContent() override;